refactor: Merge proxy test classes
Merge the proxy tests into one class.

(cherry picked from commit 95714fdb74d3cf7b9dd51d010d68e8f70881bf35)
